Update vue components
