+++ src/.vuepress/components/DownloadButtons.vue
@@ -1,25 +1,5 @@
-<template>
-	<div id="DownloadButtons">
-		<ElButton type="success" icon="el-icon-download" @click="downloadStable" @keyup.enter="downloadStable">
-			<span class="spacing">Stable</span>
-			<br />
-			<span class="downloadTag">{{ $data.tagName }}</span>
-		</ElButton>
-		<ElButton type="warning" icon="el-icon-cpu" @click="downloadPreview" @keyup.enter="downloadPreview">
-			<span class="spacing">Preview</span>
-			<br />
-			<span class="downloadTag">{{ $data.previewTagName }}</span>
-		</ElButton>
-		<span class="versionNotice">
-			Requires
-			<strong>Android 6.0</strong>
-			or higher.
-		</span>
-	</div>
-</template>
-
 <script>
-import { GITHUB_STABLE_RELEASE, GITHUB_PREVIEW_RELEASE } from "../constants";
+import { GITHUB_PREVIEW_RELEASE, GITHUB_STABLE_RELEASE } from "../constants";
 
 export default {
 	data() {
@@ -121,6 +101,26 @@ export default {
 	},
 };
 </script>
+
+<template>
+	<div id="DownloadButtons">
+		<ElButton type="success" icon="el-icon-download" @click="downloadStable" @keyup.enter="downloadStable">
+			<span class="spacing">Stable</span>
+			<br />
+			<span class="downloadTag">{{ $data.tagName }}</span>
+		</ElButton>
+		<ElButton type="warning" icon="el-icon-cpu" @click="downloadPreview" @keyup.enter="downloadPreview">
+			<span class="spacing">Preview</span>
+			<br />
+			<span class="downloadTag">{{ $data.previewTagName }}</span>
+		</ElButton>
+		<span class="versionNotice">
+			Requires
+			<strong>Android 6.0</strong>
+			or higher.
+		</span>
+	</div>
+</template>
 
 <style lang="stylus">
 #DownloadButtons
