+++ src/.vuepress/components/VersionTag.vue
@@ -1,19 +1,3 @@
-<template>
-	<span v-if="stable" class="downloadTag">{{ $data.tagName }}</span>
-	<span v-else-if="fileName" class="fileNameContainer" title="File name">
-		<MaterialIcon class="fileNameIcon" icon="get_app" />
-		<span class="fileName">aniyomi-v{{ $data.tagName }}.apk</span>
-		<slot />
-	</span>
-	<span v-else-if="preview" class="downloadTag">{{ $data.previewTagName }}</span>
-	<span v-else-if="previewFileName" class="fileNameContainer" title="File name">
-		<MaterialIcon class="fileNameIcon" icon="get_app" />
-		<span class="fileName">aniyomi-{{ $data.previewTagName }}.apk</span>
-		<slot />
-	</span>
-	<span v-else>You need to specify props.</span>
-</template>
-
 <script>
 /**
  * Code example: <VersionTag preview />
@@ -61,6 +45,22 @@ export default {
 	},
 };
 </script>
+
+<template>
+	<span v-if="stable" class="downloadTag">{{ $data.tagName }}</span>
+	<span v-else-if="fileName" class="fileNameContainer" title="File name">
+		<MaterialIcon class="fileNameIcon" icon="get_app" />
+		<span class="fileName">aniyomi-v{{ $data.tagName }}.apk</span>
+		<slot />
+	</span>
+	<span v-else-if="preview" class="downloadTag">{{ $data.previewTagName }}</span>
+	<span v-else-if="previewFileName" class="fileNameContainer" title="File name">
+		<MaterialIcon class="fileNameIcon" icon="get_app" />
+		<span class="fileName">aniyomi-{{ $data.previewTagName }}.apk</span>
+		<slot />
+	</span>
+	<span v-else>You need to specify props.</span>
+</template>
 
 <style lang="stylus" scoped>
 .fileNameContainer
