+++ src/.vuepress/components/WhatsNew.vue
@@ -1,22 +1,5 @@
-<template>
-	<div class="guide whatsNew">
-		<p class="title">
-			<MaterialIcon icon="new_releases" />
-			What's new
-		</p>
-		<!-- eslint-disable-next-line vue/no-v-html -->
-		<div v-html="whatsNew"></div>
-		<div class="note">
-			<p>
-				View the full release
-				<a href="https://github.com/jmir1/aniyomi/releases/latest" target="_blank" rel="noopener">here</a>
-			</p>
-		</div>
-	</div>
-</template>
-
 <script>
-import marked from "marked";
+import { marked } from "marked";
 
 export default {
 	data() {
@@ -38,6 +21,23 @@ export default {
 	},
 };
 </script>
+
+<template>
+	<div class="guide whatsNew">
+		<p class="title">
+			<MaterialIcon icon="new_releases" />
+			What's new
+		</p>
+		<!-- eslint-disable-next-line vue/no-v-html -->
+		<div v-html="whatsNew"></div>
+		<div class="note">
+			<p>
+				View the full release
+				<a href="https://github.com/jmir1/aniyomi/releases/latest" target="_blank" rel="noopener">here</a>
+			</p>
+		</div>
+	</div>
+</template>
 
 <style lang="stylus">
 .whatsNew
