+++ src/.vuepress/theme/components/NavBar.vue
@@ -1,46 +1,11 @@
-<template>
-	<header class="navbar">
-		<!-- eslint-disable-next-line vue/require-explicit-emits -->
-		<SidebarButton @toggle-sidebar="$emit('toggle-sidebar')" />
-
-		<RouterLink :to="$localePath" class="home-link">
-			<img
-				v-if="$site.themeConfig.logo"
-				class="logo"
-				:src="$withBase($site.themeConfig.logo)"
-				:alt="$siteTitle"
-			/>
-			<span v-if="$siteTitle" ref="siteName" class="site-name" :class="{ 'can-hide': $site.themeConfig.logo }">
-				{{ $siteTitle }}
-			</span>
-		</RouterLink>
-
-		<div
-			class="links"
-			:style="
-				linksWrapMaxWidth
-					? {
-							'max-width': linksWrapMaxWidth + 'px',
-					  }
-					: {}
-			"
-		>
-			<div v-if="$page.frontmatter.hideSearch" />
-			<AlgoliaSearchBox v-else-if="isAlgoliaSearch" :options="algolia" />
-			<SearchBox v-else-if="$site.themeConfig.search !== false && $page.frontmatter.search !== false" />
-			<NavLinks class="can-hide" />
-		</div>
-	</header>
-</template>
-
 <script>
 import SearchBox from "@SearchBox";
-import SidebarButton from "@theme/components/SidebarButton.vue";
 import NavLinks from "@theme/components/NavLinks.vue";
+import SidebarButton from "@theme/components/SidebarButton.vue";
 import AlgoliaSearchBox from "./AlgoliaSearchBox.vue";
 
 export default {
-	name: "Navbar",
+	name: "NavBar",
 
 	components: {
 		SidebarButton,
@@ -95,6 +60,41 @@ function css(el, property) {
 }
 </script>
 
+<template>
+	<header class="navbar">
+		<!-- eslint-disable-next-line vue/require-explicit-emits -->
+		<SidebarButton @toggle-sidebar="$emit('toggle-sidebar')" />
+
+		<RouterLink :to="$localePath" class="home-link">
+			<img
+				v-if="$site.themeConfig.logo"
+				class="logo"
+				:src="$withBase($site.themeConfig.logo)"
+				:alt="$siteTitle"
+			/>
+			<span v-if="$siteTitle" ref="siteName" class="site-name" :class="{ 'can-hide': $site.themeConfig.logo }">
+				{{ $siteTitle }}
+			</span>
+		</RouterLink>
+
+		<div
+			class="links"
+			:style="
+				linksWrapMaxWidth
+					? {
+							'max-width': linksWrapMaxWidth + 'px',
+					  }
+					: {}
+			"
+		>
+			<div v-if="$page.frontmatter.hideSearch" />
+			<AlgoliaSearchBox v-else-if="isAlgoliaSearch" :options="algolia" />
+			<SearchBox v-else-if="$site.themeConfig.search !== false && $page.frontmatter.search !== false" />
+			<NavLinks class="can-hide" />
+		</div>
+	</header>
+</template>
+
 <style lang="stylus">
 $navbar-vertical-padding = 0.7rem
 $navbar-horizontal-padding = 1.5rem
@@ -136,7 +136,6 @@ $navbar-horizontal-padding = 1.5rem
 		.links
 			padding-left 1.5rem
 		.site-name
-			width calc(100vw - 9.4rem)
 			overflow hidden
 			white-space nowrap
 			text-overflow ellipsis
